[TASK] HTML Parser
Currently, the DeepL write API can't handle HTML inside the request.
For dealing with RTE configured fields and handle these correct, a basic
HTML parser is added.

This is just a proof of concept and has many lags at the moment, e.g.
not supported inline tags like a or span. This has to be adjusted.