[TASK] Refactor Langauge Support enum to class
In preparation of getting the respective values via DeepL API, the Enum
for supported languages is refactored as a normal class with constants.

At the moment, there is no description of how the DeepL API will
implement the `/language` endpoint, so this implementation is just to be
seen as Proof of Concept.