[TASK] Remove obsolete TypeScript files
For the extension kickoff deepltranslate-core has been used
as skeleton, containing some TypeScript files to build TYPO3
backend JavaScript overrides (localization modal).

It was not suitable to build that in a cross extension matrix
version with `deepltranslate-core` and has been moved into a
dedicated extension `deepl-base` doing the override and also
providing `PSR-14` events to allow extension adding custom
to the modal.

That makes them here obsolete and are now removed to keep the
house clean.