[BUGFIX] Adjust HTML parser
The `HtmlParser` used the DOM Document methods saving the html for the
return to the system. THis is fine, but with `DOMDocument::saveHTML()`
all possible characters were converted to their respective html
entities. While working inside TYPO3, the main character encoding is
`UTF-8` in most cases and the result should return the content `as-is`
and not with entities encoded.

With switching the parser to return the rendered XML, this issue doesn't
appear, as `DOMDocument::saveXML()` doesn't convert to HTML entities.